activate the new structural search execution path
